+++ OUDS/Core/Components/Sources/Navigations/ToolBar/OUDSToolBarBottom.swift
@@ -36,11 +36,11 @@ import SwiftUI
 ///
 /// ## Badges uses
 ///
-/// Because the system does not render system badge for items in bottom toolbar, the ``OUDSBadge`` is used for iOS 26 with Liquid Glass and bottom toolbars.
+/// Because the system does not render system badge for items in bottom toolbar, the OUDS badge component is used for iOS 26 with Liquid Glass and bottom toolbars.
 /// However the glassified effect of Liquid Glass does not make sometimes the badge readable, that is the reason why you should use badge for bottom toolbars with lots of care
 /// and prefer *prominent* or *tinted* styles for the item with the badge instead of *default*.
 ///
-/// For iOS until 18 and without Liquid Glass ``OUDSBadge`` is always used.
+/// For iOS until 18 and without Liquid Glass the OUDS badge components are always used.
 ///
 /// ## Code samples
 ///